# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the NutriMind nutrition and health tracking mobile app.",
        "title": "NutriMind API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
        "/ai/advice": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns personalised nutrition advice from the AI coach based on today's logged calories and water intake vs. the user's targets. An optional custom prompt can steer the advice. The response always includes a disclaimer field.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "Get AI nutrition advice",
                "parameters": [
                    {
                        "description": "Optional prompt (body may be omitted entirely)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAdviceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AI advice with context summary",
                        "schema": {
                            "$ref": "#/definitions/dto.AdviceResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "ONBOARDING_REQUIRED",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error or AI service error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ai/meal-suggestion": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a meal suggestion for a given meal type (breakfast, lunch, dinner, snack) tailored to the user's calorie and macro targets. Estimated nutritional values are included.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "Get AI meal suggestion",
                "parameters": [
                    {
                        "description": "Meal type to suggest for",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetMealSuggestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suggested meal with estimated macros and disclaimer",
                        "schema": {
                            "$ref": "#/definitions/dto.MealSuggestionResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error (invalid meal_type)",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "ONBOARDING_REQUIRED",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error or AI service error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AdviceContextSummaryResponse": {
            "type": "object",
            "properties": {
                "calorie_target": {
                    "type": "number"
                },
                "calories_logged": {
                    "type": "number"
                },
                "water_ml_logged": {
                    "type": "integer"
                },
                "water_target_ml": {
                    "type": "integer"
                }
            }
        },
        "dto.AdviceResponse": {
            "type": "object",
            "properties": {
                "advice": {
                    "type": "string"
                },
                "context_summary": {
                    "$ref": "#/definitions/dto.AdviceContextSummaryResponse"
                },
                "disclaimer": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetAdviceRequest": {
            "type": "object",
            "properties": {
                "prompt": {
                    "description": "optional",
                    "type": "string"
                }
            }
        },
...
        "dto.GetMealSuggestionRequest": {
            "type": "object",
            "required": [
                "meal_type"
            ],
            "properties": {
                "meal_type": {
                    "type": "string",
                    "enum": [
                        "BREAKFAST",
                        "LUNCH",
                        "DINNER",
                        "SNACK"
                    ]
                }
            }
        },
...
        "dto.MealSuggestionResponse": {
            "type": "object",
            "properties": {
                "disclaimer": {
                    "type": "string"
                },
                "estimated_calories": {
                    "type": "number"
                },
                "estimated_carb_g": {
                    "type": "number"
                },
                "estimated_fat_g": {
                    "type": "number"
                },
                "estimated_protein_g": {
                    "type": "number"
                },
                "suggestion": {
                    "type": "string"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.Error"
                },
                "is_success": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token — format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}